refactor: move rpc/error code to a new pkg
